+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-form-web/bnd.bnd
@@ -1,6 +1,7 @@
 Bundle-Name: Liferay Dynamic Data Lists Form Web
 Bundle-SymbolicName: com.liferay.dynamic.data.lists.form.web
-Bundle-Version: 1.0.12
+Bundle-Version: 1.1.0
+Export-Package: com.liferay.dynamic.data.lists.form.web.constants;
 Liferay-JS-Config: /META-INF/resources/admin/js/config.js
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title: Dynamic Data Lists
