+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-form-web/src/main/java/com/liferay/dynamic/data/lists/form/web/constants/DDLFormWebKeys.java
@@ -12,17 +12,14 @@
  * details.
  */
 
-package com.liferay.dynamic.data.lists.form.web.internal.constants;
+package com.liferay.dynamic.data.lists.form.web.constants;
 
 /**
  * @author Marcellus Tavares
  */
-public class DDLFormPortletKeys {
+public class DDLFormWebKeys {
 
-	public static final String DYNAMIC_DATA_LISTS_FORM =
-		"com_liferay_dynamic_data_lists_form_web_portlet_DDLFormPortlet";
-
-	public static final String DYNAMIC_DATA_LISTS_FORM_ADMIN =
-		"com_liferay_dynamic_data_lists_form_web_portlet_DDLFormAdminPortlet";
+	public static final String DYNAMIC_DATA_LISTS_RECORD_SET =
+		"DYNAMIC_DATA_LISTS_RECORD_SET";
 
 }
