+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-form-web/src/main/resources/com/liferay/dynamic/data/lists/form/web/constants/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
